simple component to share same header on slideover
