+++ frontend/app/src/screens/groups/add-group-trigger-button.tsx
@@ -1,16 +1,12 @@
 import { ButtonWithTooltip } from "@/components/buttons/button-primitive";
 import { usePermission } from "@/hooks/usePermission";
-import { useAtomValue } from "jotai";
-import { currentBranchAtom } from "@/state/atoms/branches.atom";
 import { useState } from "react";
-import SlideOver from "@/components/display/slide-over";
+import SlideOver, { SlideOverTitle } from "@/components/display/slide-over";
 import { Icon } from "@iconify-icon/react";
 import AddGroupForm from "@/screens/groups/add-group-form";
 import { iNodeSchema } from "@/state/atoms/schema.atom";
 import graphqlClient from "@/graphql/graphqlClientApollo";
 import { useObjectDetails } from "@/hooks/useObjectDetails";
-import { Badge } from "@/components/ui/badge";
-import { ObjectHelpButton } from "@/components/menu/object-help-button";
 
 type AddGroupTriggerButtonProps = {
   schema: iNodeSchema;
@@ -23,7 +19,6 @@ export default function AddGroupTriggerButton({
   ...props
 }: AddGroupTriggerButtonProps) {
   const permission = usePermission();
-  const currentBranch = useAtomValue(currentBranchAtom);
   const [isAddGroupFormOpen, setIsAddGroupFormOpen] = useState(false);
 
   const { data } = useObjectDetails(schema, objectId);
@@ -46,35 +41,12 @@ export default function AddGroupTriggerButton({
       <SlideOver
         offset={1}
         title={
-          <div className="space-y-2">
-            <div className="flex">
-              <Badge variant="blue" className="flex items-center gap-1">
-                <Icon icon="mdi:layers-triple" />
-                <span>{currentBranch?.name}</span>
-              </Badge>
-
-              <ObjectHelpButton
-                kind={schema.kind}
-                documentationUrl={schema.documentation}
-                className="ml-auto"
-              />
-            </div>
-
-            <div className="flex justify-between">
-              <div className="text-sm flex items-center gap-2 whitespace-nowrap">
-                {schema.label}
-
-                <Icon icon="mdi:chevron-right" />
-
-                <span className="truncate">{objectDetailsData?.display_label}</span>
-              </div>
-            </div>
-
-            <div>
-              <h3 className="text-lg font-semibold">Select group(s)</h3>
-              <p className="text-sm">Select one or more groups to assign</p>
-            </div>
-          </div>
+          <SlideOverTitle
+            schema={schema}
+            currentObjectLabel={objectDetailsData?.display_label}
+            title="Select group(s)"
+            subtitle="Select one or more groups to assign"
+          />
         }
         open={isAddGroupFormOpen}
         setOpen={setIsAddGroupFormOpen}>
